final updates before first delivery

+++ bin/clean.sh
@@ -7,4 +7,4 @@ DIR=$(cd $(dirname "${BASH_SOURCE[0]}") && pwd)
 
 . $DIR/common.sh
 
-find "$DATA_DIR" -mindepth 1 -delete
+test -d "$DATA_DIR" && find "$DATA_DIR" -mindepth 1 -delete || true
